requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
shap>=0.43.0
mysql-connector-python>=8.0.0  # ¡CAMBIADO!
python-dotenv>=1.0.0
bcrypt>=4.0.0
streamlit-authenticator>=0.2.0
streamlit-option-menu>=0.3.0
watchdog>=3.0.0
sqlalchemy>=2.0.0
pymysql>=1.0.0